Add ANSI color parsing and expose Colored enum.
This adds `parse_ansi` functions to the Colored and Color enums. These
try to parse ANSI color sequences which appear in places such as
/etc/os-release and other configuration files.

Tests are provided using the formatter for Color and Colored enums.

This doesn't include support the the 3/4 bit ansi colors mainly because
I'm unsure about how they should be handled.