windows: {enable,disable}_raw_mode on CONIN$
On Windows, `enable_raw_mode` and `disable_raw_mode` were attempting to
read the mode from `Handle::input_handle()`, which resolves to the
result of `GetStdHandle(STD_INPUT_HANDLE)`. When stdin is not piped,
this works fine because that call will resolve to `CONIN$`. However,
when stdin is piped, the call will fail because `GetConsoleMode()` only
operates on Console handles, not pipes.

Changing this to always operate on `Handle::current_in_handle()`, which
always resolves to `CONIN$`, results in being able to enable/disable raw
mode on the console. This allows for using crossterm on Windows when
stdin is a pipe.

Fixes #452