Handle key events that don't provide a character on Windows
Many key combinations produce key events which have u_char == 0, and these
have been discarded until now. This for example includes all combinations
involving the Ctrl+Alt modifiers, as well as many key combinations with just
Ctrl. We can provide events for such key combinations by determining the
character associated with the keys from consulting the keyboard layout. Almost
all keys on a keyboard have characters associated with them -- it's just a
question of whether we can determine what character corresponds to a key event.
There are some caveats involved in doing that...

In addition, the key events with u_char in the ASCII control code range was
until now mapped into the ASCII range '@' to '_' which is inaccurate for many
keys and for users with non-US keyboard layouts. The character for key
combinations that produce control codes are now also handled by consulting the
keyboard layout.

The caveats revolve around determining the keyboard layout, which has two
issues:
 1. There is a race condition between the user typing in their terminal with one
    keyboard layout active, and the console application determining the keyboard
    layout while processing the key event later, as these two events happen
    asynchronously. If a user changes the active keyboard layout in between the
    two events, then the console application might misinterpret the character.
 2. For console applications running in a Conhost terminal, it turns out to be
    very difficult to determine the active keyboard layout. There appears to be
    no available APIs that reliably provide the layout.