feat: add synchronized output/update
The Contour terminal describes a VT extension for pausing rendering in
the terminal during updates, preventing tearing and other artifacts.
This extension is already supported by several terminals, with support
planned for many others.

See: https://gist.github.com/christianparpart/d8a62cc1ab659194337d73e399004036

This adds the two commands related to pausing and resuming rendering:
`BeginSynchronizedUpdate` and `EndSynchronizedUpdate`. It also adds a
utility function that performs a function within a synchronized update.

A demo of this is also provided in the interactive example, that slowly
renders a list of numbers - once without synchronized updates, and then
once with.