terminal:size:handle window size with zero rows or columns
This commit handles the case where window_size returns success, but the
number of rows and/or columns is 0. Instead of returning Ok((0,0)), the
size function will use tput as the second source of sizes.

In practice, window_size returning zero rows or columns is rare when
the terminal is running over SSH, desktop, etc. However, it occurs
frequently when the terminal is running over serial lines, especially on
embedded devices with minimal environments. (And this is how the problem
was detected)